Support basic templates in YAML files
A little bit inspired by Ansible. The reason for this feature is that
in practice DAG tasks often are quite similar a lot but get different
on few details. To not let users write a lot of copy-pasted text we
provide very basic template system to describe recurrent things.

Basically, `do...with items` is a `for`-loop which generates body with
a given list of items and merges it with the rest DAG fields.

The key moment here is that after template render you should get valid
YAML which could be used as a source for airflow-declarative parser
without additional modifications. This should help to understand into
what template will get rendered and simplify dags processing by the rest
systems which knows nothing about this feature.