Fix `with_items` raising TypeError for jinja like `{{ 1 + item.value }}`
For such templates there was an exception:
   TypeError: unsupported operand type(s) for +: 'int' and 'str'

The reason is that we implicitly append `| yaml` to jinja templates
in `with_items`, so the template starts to look like `{{ 1 + item.value | yaml }}`,
which is effectively an "int + str" expression.

This commit fixes that by implicitly wrapping all templates in parens,
so the aforementioned template would become `{{ (1 + item.value) | yaml }}`.